Ajuste de ventas

+++ src/main/java/co/com/hotelyando/core/model/SaleReport.java
@@ -62,11 +62,11 @@ public class SaleReport implements Serializable{
 	@Data
 	public static class Values {
 		
-		private Integer gross;
-		private Integer tax;
-		private Integer net;
-		private Integer discount;
-		private Integer total;
+		private Double gross;
+		private Double tax;
+		private Double net;
+		private Double discount;
+		private Double total;
 	}
 
 }
